weather-api/requirements.txt
fastapi==0.115.0
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.31
